refactoring: domainディレクトリの中に、modelを入れる
