+++ api/usecase/like.go
@@ -2,7 +2,7 @@ package usecase
 
 import (
 	"github.com/htoyoda18/TweetAppV2/api/controller/handler/request"
-	"github.com/htoyoda18/TweetAppV2/api/model"
+	"github.com/htoyoda18/TweetAppV2/api/domain/model"
 	"github.com/htoyoda18/TweetAppV2/api/repository"
 	"github.com/htoyoda18/TweetAppV2/api/shared"
 	"gorm.io/gorm"
